+++ components/(dashboards)/receipts/extraction/(tabs)/edit/ReceiptDataForm.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { UseFormReturn } from 'react-hook-form';
 
 import ExtractionFormComponent from '@/components/(dashboards)/bills/extraction/components/ExtractionFormComponent';
-import Approvals from '@/components/(shared)/invoices/Approvals';
+import Approvals from '@/components/(shared)/Approvals';
 import Container from '@/components/ui/container';
 import LoadingState from '@/components/ui/empty-state';
 import { Form } from '@/components/ui/form';
